requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.4
django-debug==0.1.1
docopt==0.6.2
idna==2.6
Pillow==5.1.0
pipreq==0.4
pipreqs==0.4.9
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.3
requests==2.18.4
urllib3==1.22
yarg==0.1.9
django-debug-toolbar==1.9.1